Setup Engine to append request to event payload
In order to log additional request-level details, we need to hook into
ApplicationController. To hook into this controller, we need to wire up
a Rails engine to process our decorator.

This change also moves the testing of the controller to
ApplicationController instead of Spree::Admin::BaseController. This was
more appropriate now given this level of integration.